+++ CHANGELOG.md
@@ -8,8 +8,12 @@ Changes in the original version can be found from CHANGELOG.old.md file
 - Updated docusaur to latest 2.x version and published docs in github pages
 - Fixed failing unit test and added node versions 20,22,24 to pipeline test matrix
 - Bump eslint from 8 to 9
+- Remove yarn from project
 - Convert unit tests to typescript (+ bump jest from 29 to 30)
+- Drop dependencies: npm-run-all and lodash.mergewith
 - Dependency updates
+- Added dev container support
+- Breaking change: swaggerJsdoc() now returns a promise
 
 ## [7.0.0] - version skipped